% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isimip-package.R
\docType{package}
\name{isimip-package}
\alias{isimip}
\alias{isimip-package}
\title{isimip: Client for the ISIMIP Repository API}
\description{
A thin client library to use the API of the ISIMIP repository (https://data.isimip.org) using R. This package provides methods to query datasets and files, submit processing jobs (cutout, mask, select operations), and download results from the ISIMIP climate impact data repository.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/limnotrack/isimip-client}
  \item \url{https://limnotrack.github.io/isimip-client/}
  \item Report bugs at \url{https://github.com/limnotrack/isimip-client/issues}
}

}
\author{
\strong{Maintainer}: Tadhg Moore \email{tadhg.moore@example.com}

Authors:
\itemize{
  \item Jochen Klar (Original Python implementation)
}

}
\keyword{internal}
