% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ISIMIPClient}
\alias{ISIMIPClient}
\title{Create a new ISIMIP client}
\usage{
ISIMIPClient(
  data_url = "https://data.isimip.org/api/v1",
  files_api_url = "https://files.isimip.org/api/v2",
  files_api_version = "v2",
  auth = NULL,
  headers = NULL
)
}
\arguments{
\item{data_url}{Base URL for the ISIMIP data API}

\item{files_api_url}{Base URL for the ISIMIP files API}

\item{files_api_version}{Version of the files API ("v1" or "v2")}

\item{auth}{Optional authentication (list with username and password)}

\item{headers}{Optional additional HTTP headers}
}
\value{
A new \code{ISIMIPClient} object
}
\description{
Create a new ISIMIP client
}
